Datasource Cleanup
  - Moves flatfile-dependant methods from TownySQLSource and
TownyFlatfileSource to TownyDatabaseHandler.
    - PlotBlockData/RegenList/SnapshotList methods.
  - Removes TownySQLSource's ffQueryQueue.
  - Moves TownyFlatfileSource queryQueue to TownyDatabaseHandler.
  - Moves methods around to put things back into a sensible order.

  - Fix a call to getTownWorld() that should have been getWorld() in
TownyPlayerListener.

  - Use previously unused TownyDatasource method to load residents'
friends.
  - Use previously unused TownyDatasource method to load nations' allies
and enemies.

  - Deprecates getTownWorld().
  - Deprecates getResidentKeys(). (Unused)
  - Deprecates getTownsKeys(). (Unused)
  - Deprecates getNationsKeys(). (Unused)
